auth_service/requirements.txt
flask
python-keycloak
requests
gunicorn
Flask-JWT-Extended 
python-dotenv